add naming scheme {:04}.* to data generation fct, also refactored responsibilities
