+++ OpenGLSnake/Application.cpp
@@ -101,9 +101,11 @@ void Application::openGLRender(const std::string_view board) {
 }
 
 void Application::genData() {
+    static int count {};
     if (prev_board_state != board_state) {
         prev_board_state = board_state;
-        m_ascii_renderer.generateData();
-        m_renderer->generateData();
+        m_ascii_renderer.generateData("in", count);
+        m_renderer->generateData("out", count);
+        ++count;
     }
 }
