+++ OpenGLSnake/Renderer/RenderTexture.cpp
@@ -1,10 +1,8 @@
 #include <glad/glad.h>
 
 #include "RenderTexture.h"
-#include "../assets.h"
+#include "../settings.h"
 
-#define STB_IMAGE_WRITE_IMPLEMENTATION
-#include "../extern/stb_image_write.h"
 
 RenderTexture::RenderTexture(const int size)
     : m_size{size} {
@@ -35,22 +33,9 @@ void RenderTexture::end(const int width, const int height) {
     glViewport(0, 0, width, height);
 }
 
-void saveTexAsImg(const std::string& filename, const int w, const int h, const std::vector<unsigned char>& rgba) {
-    // Flip vertically for correct orientation
-    std::vector<unsigned char> flipped(w * h * 4);
-    for (int y = 0; y < h; ++y) {
-        memcpy(&flipped[y * w * 4],
-               &rgba[(h - 1 - y) * w * 4],
-               w * 4);
-    }
-
-    stbi_write_png(filename.c_str(), w, h, 4, flipped.data(), w*4);
-}
-
-void RenderTexture::getTextureImage() {
+const std::vector<unsigned char>& RenderTexture::getTextureImage() {
     glGetTextureImage(m_Tex, 0, GL_RGBA, GL_UNSIGNED_BYTE, pixels.size(), pixels.data());
-    static int count {};
-    saveTexAsImg(data::path(std::to_string(count++)+".png"), 32, 32, pixels);
+    return pixels;
 }
 
 void RenderTexture::create() {
